qcu/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qcu"
version = "0.2.0"
description = "QCU — Phase-based quantum coprocessor subsystem"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
]

[project.optional-dependencies]
gpu = [
    "cupy-cuda13x>=13.0",   # CUDA 13.x；勿安装 cupy-cuda12x
]
qiskit = [
    "qiskit>=1.0",
]
qasm3 = [
    "openqasm3[parser]>=0.5",
]
cirq = [
    "cirq-core>=1.3",
]
all = [
    "cupy-cuda13x>=13.0",
    "qiskit>=1.0",
    "openqasm3[parser]>=0.5",
    "cirq-core>=1.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["qcu*", "qcu_lang*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json"]
